Island: Rename password_based_bytes_encrypt{ion,or}.py
